1.染色单、采购单与其他单据的 打印纸张不同，尺寸要重新写 2.纱线、材料的数量均改为 小数，现在都是整数 3.生产单 -》生产材料信息 -》 数量改为 【数量（kg）】 4.染色进度单、染色单、辅料采购单、原材料采购单 -》 纱线信息 -》 数量后加（kg） 5.原材料记录单 -》 材料列表 -》总数量后加（kg） 6.车缝记录单  去掉 【生产单位】 7. 整烫记录单 去掉【生产单位】 8.单据所有公司显示简称 9.系统信息管理 -- 业务员，显示公司，且按照公司ID进行排序

+++ src/com/fuwei/entity/ordergrid/ColoringOrderDetail.java
@@ -3,7 +3,7 @@ package com.fuwei.entity.ordergrid;
 public class ColoringOrderDetail {
 	private String color;//色号
 	private Integer material ; //材料
-	private int quantity;//数量
+	private double quantity;//数量
 	private String standardyarn;//标准样纱
 	public Integer getMaterial() {
 		return material;
@@ -11,10 +11,10 @@ public class ColoringOrderDetail {
 	public void setMaterial(Integer material) {
 		this.material = material;
 	}
-	public int getQuantity() {
+	public double getQuantity() {
 		return quantity;
 	}
-	public void setQuantity(int quantity) {
+	public void setQuantity(double quantity) {
 		this.quantity = quantity;
 	}
 	public String getColor() {
